requirements.txt
librosa==0.10.2.post1
numpy==1.26.1
mutagen==1.47.0
pillow==11.0.0
Flask==3.1.0
Flask-Cors==5.0.0 
python-dotenv==1.0.1